Move common functions and macros from num2&num3 to num_common
